# packages/webui/frontend/src/api/openapi.snapshot.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "linling-webui",
    "version": "0.0.0"
  },
  "paths": {
...
    "/api/rules": {
      "get": {
        "tags": [
          "rules"
        ],
        "summary": "List Rules",
        "operationId": "list_rules_api_rules_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RuleSummary"
                  },
                  "title": "Response List Rules Api Rules Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/rules/{name}/hits": {
      "get": {
        "tags": [
          "rules"
        ],
        "summary": "Hits For",
        "description": "Return recent hits for a given handler name.\n\nTime is emitted as ISO-8601 (UTC) to match ``/api/audit`` and\n``/api/events``; the UI should not have to switch formatters per\nendpoint.",
        "operationId": "hits_for_api_rules__name__hits_get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response Hits For Api Rules  Name  Hits Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/rules/files": {
      "get": {
        "tags": [
          "rules"
        ],
        "summary": "List Rule Files",
        "operationId": "list_rule_files_api_rules_files_get",
        "parameters": [
          {
            "name": "bot_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Bot Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RuleFile"
                  },
                  "title": "Response List Rule Files Api Rules Files Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/rules/files/content": {
      "get": {
        "tags": [
          "rules"
        ],
        "summary": "Read Rule File",
        "operationId": "read_rule_file_api_rules_files_content_get",
        "parameters": [
          {
            "name": "bot_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Bot Id"
            }
          },
          {
            "name": "path",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Bot-relative file path",
              "title": "Path"
            },
            "description": "Bot-relative file path"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleFileContent"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "rules"
        ],
        "summary": "Save Rule File",
        "description": "Save a ``.ling`` file. Requires ``bot_admin`` — saves can hot-reload\nthe live router which would otherwise let any read-only viewer\nrewrite handler logic.",
        "operationId": "save_rule_file_api_rules_files_content_put",
        "parameters": [
          {
            "name": "bot_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Bot Id"
            }
          },
          {
            "name": "path",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Path"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuleFileSaveRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleFileSaveResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/rules/lint": {
      "post": {
        "tags": [
          "rules"
        ],
        "summary": "Lint Rule Source",
        "description": "Lint a raw ``.ling`` snippet; does not touch disk.\n\nWe accept the raw content (rather than a bot/path pair) so the\neditor can live-lint without each keystroke requiring a file save.\nNo bot visibility check because no data crosses tenant boundaries.",
        "operationId": "lint_rule_source_api_rules_lint_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Req"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleLintResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RuleFile": {
        "properties": {
          "path": {
            "type": "string",
            "title": "Path"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          },
          "handler_count": {
            "type": "integer",
            "title": "Handler Count",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "path",
          "size"
        ],
        "title": "RuleFile",
        "description": "Metadata entry for one ``.ling`` rule file."
      },
      "RuleFileContent": {
        "properties": {
          "path": {
            "type": "string",
            "title": "Path"
          },
          "content": {
            "type": "string",
            "title": "Content"
          }
        },
        "type": "object",
        "required": [
          "path",
          "content"
        ],
        "title": "RuleFileContent",
        "description": "Full content of one rule file."
      },
      "RuleFileSaveRequest": {
        "properties": {
          "content": {
            "type": "string",
            "title": "Content"
          },
          "reload": {
            "type": "boolean",
            "title": "Reload",
            "default": true
          },
          "lint_first": {
            "type": "boolean",
            "title": "Lint First",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "content"
        ],
        "title": "RuleFileSaveRequest",
        "description": "Body for updating a rule file."
      },
      "RuleFileSaveResult": {
        "properties": {
          "saved": {
            "type": "boolean",
            "title": "Saved"
          },
          "issues": {
            "items": {
              "$ref": "#/components/schemas/RuleLintIssue"
            },
            "type": "array",
            "title": "Issues"
          },
          "reloaded": {
            "type": "boolean",
            "title": "Reloaded",
            "default": false
          },
          "handlers": {
            "type": "integer",
            "title": "Handlers",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "saved"
        ],
        "title": "RuleFileSaveResult",
        "description": "Response from a save+reload cycle."
      },
      "RuleLintIssue": {
        "properties": {
          "line": {
            "type": "integer",
            "title": "Line"
          },
          "col": {
            "type": "integer",
            "title": "Col",
            "default": 0
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "severity": {
            "type": "string",
            "enum": [
              "error",
              "warning",
              "info"
            ],
            "title": "Severity"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "line",
          "code",
          "severity",
          "message"
        ],
        "title": "RuleLintIssue",
        "description": "One lint finding."
      },
      "RuleLintResult": {
        "properties": {
          "issues": {
            "items": {
              "$ref": "#/components/schemas/RuleLintIssue"
            },
            "type": "array",
            "title": "Issues"
          },
          "handler_count": {
            "type": "integer",
            "title": "Handler Count",
            "default": 0
          }
        },
        "type": "object",
        "title": "RuleLintResult",
        "description": "Response from linting a snippet."
      },
      "RuleSummary": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "trigger": {
            "type": "string",
            "title": "Trigger"
          },
          "hits_today": {
            "type": "integer",
            "title": "Hits Today",
            "default": 0
          },
          "avg_latency_ms": {
            "type": "number",
            "title": "Avg Latency Ms",
            "default": 0
          },
          "last_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error"
          }
        },
        "type": "object",
        "required": [
          "name",
          "trigger"
        ],
        "title": "RuleSummary"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}